move defaults to pipeline and use in component decorator
