+++ spacy/pipeline/defaults/__init__.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from .... import util
+from spacy import util
 
 
 def default_nel_config():
