+++ spacy/pipeline/defaults/entity_linker_defaults.cfg
