+++ spacy/pipeline/defaults/morphologizer_defaults.cfg
