+++ spacy/pipeline/defaults/ner_defaults.cfg
