+++ spacy/pipeline/defaults/parser_defaults.cfg
