+++ spacy/pipeline/defaults/senter_defaults.cfg
