+++ spacy/pipeline/defaults/tagger_defaults.cfg
