+++ spacy/pipeline/defaults/tensorizer_defaults.cfg
