+++ spacy/pipeline/defaults/textcat_bow_defaults.cfg
