+++ spacy/pipeline/defaults/textcat_cnn_defaults.cfg
