+++ spacy/pipeline/defaults/textcat_defaults.cfg
