+++ spacy/pipeline/defaults/tok2vec_defaults.cfg
