+++ spacy/pipeline/pipes.pyx
@@ -2,6 +2,10 @@
 import numpy
 import srsly
 import random
+
+from spacy.pipeline.defaults import default_tagger, default_parser,  default_ner,  default_textcat
+from spacy.pipeline.defaults import default_nel, default_senter, default_tensorizer
+
 from thinc.api import CosineDistance, to_categorical, get_array_module
 from thinc.api import set_dropout_rate
 import warnings
@@ -234,7 +238,7 @@ class Pipe(object):
         return self
 
 
-@component("tensorizer", assigns=["doc.tensor"])
+@component("tensorizer", assigns=["doc.tensor"], default_model=default_tensorizer)
 class Tensorizer(Pipe):
     """Pre-train position-sensitive vectors for tokens."""
 
@@ -366,7 +370,7 @@ class Tensorizer(Pipe):
         return sgd
 
 
-@component("tagger", assigns=["token.tag", "token.pos", "token.lemma"])
+@component("tagger", assigns=["token.tag", "token.pos", "token.lemma"], default_model=default_tagger)
 class Tagger(Pipe):
     """Pipeline component for part-of-speech tagging.
 
@@ -651,7 +655,7 @@ class Tagger(Pipe):
         return self
 
 
-@component("senter", assigns=["token.is_sent_start"])
+@component("senter", assigns=["token.is_sent_start"], default_model=default_senter)
 class SentenceRecognizer(Tagger):
     """Pipeline component for sentence segmentation.
 
@@ -991,7 +995,7 @@ class ClozeMultitask(Pipe):
             losses[self.name] += loss
 
 
-@component("textcat", assigns=["doc.cats"])
+@component("textcat", assigns=["doc.cats"], default_model=default_textcat)
 class TextCategorizer(Pipe):
     """Pipeline component for text classification.
 
@@ -1242,7 +1246,8 @@ cdef class EntityRecognizer(Parser):
 @component(
     "entity_linker",
     requires=["doc.ents", "doc.sents", "token.ent_iob", "token.ent_type"],
-    assigns=["token.ent_kb_id"]
+    assigns=["token.ent_kb_id"],
+    default_model=default_nel,
 )
 class EntityLinker(Pipe):
     """Pipeline component for named entity linking.
@@ -1688,8 +1693,19 @@ class Sentencizer(Pipe):
 
 
 # Cython classes can't be decorated, so we need to add the factories here
-Language.factories["parser"] = lambda nlp, model, **cfg: DependencyParser.from_nlp(nlp, model, **cfg)
-Language.factories["ner"] = lambda nlp, model, **cfg: EntityRecognizer.from_nlp(nlp, model, **cfg)
+Language.factories["parser"] = lambda nlp, model, **cfg: parser_factory(nlp, model, **cfg)
+Language.factories["ner"] = lambda nlp, model, **cfg: ner_factory(nlp, model, **cfg)
 
+def parser_factory(nlp, model, **cfg):
+    if model is None:
+        model = default_parser()
+        warnings.warn(Warnings.W098.format(name="parser"))
+    return DependencyParser.from_nlp(nlp, model, **cfg)
+
+def ner_factory(nlp, model, **cfg):
+    if model is None:
+        model = default_ner()
+        warnings.warn(Warnings.W098.format(name="ner"))
+    return EntityRecognizer.from_nlp(nlp, model, **cfg)
 
 __all__ = ["Tagger", "DependencyParser", "EntityRecognizer", "Tensorizer", "TextCategorizer", "EntityLinker", "Sentencizer", "SentenceRecognizer"]
