+++ spacy/tests/parser/test_ner.py
@@ -2,7 +2,7 @@ import pytest
 
 from spacy import util
 from spacy.lang.en import English
-from spacy.ml.models.defaults import default_ner
+from spacy.pipeline.defaults import default_ner
 
 from spacy.pipeline import EntityRecognizer, EntityRuler
 from spacy.vocab import Vocab
