+++ spacy/tests/parser/test_preset_sbd.py
@@ -4,7 +4,7 @@ from spacy.attrs import NORM
 from spacy.gold import GoldParse
 from spacy.vocab import Vocab
 
-from spacy.ml.models.defaults import default_parser
+from spacy.pipeline.defaults import default_parser
 from spacy.tokens import Doc
 from spacy.pipeline import DependencyParser
