+++ spacy/tests/regression/test_issue4042.py
@@ -3,7 +3,7 @@ from spacy.pipeline import EntityRecognizer, EntityRuler
 from spacy.lang.en import English
 from spacy.tokens import Span
 from spacy.util import ensure_path
-from spacy.ml.models.defaults import default_ner
+from spacy.pipeline.defaults import default_ner
 
 from ..util import make_tempdir
